% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_to_edgelist.R
\name{adjacency.to.edgelist}
\alias{adjacency.to.edgelist}
\title{adjacency.to.edgelist}
\usage{
adjacency.to.edgelist(adjacency, mode = c("undirected", "directed"),
  weighted = NULL)
}
\arguments{
\item{adjacency:}{a list object whose tth entry is an adjacency matrix representing the tth layer of a multilayer network.}

\item{mode:}{directed or undirected}

\item{weighted:}{currently not functioning. Coming in later version.}
}
\value{
edgelist: a matrix with three columns representing edge connections- node1, node2, layer
}
\description{
Function that converts a list of adjacency matrices to an edgelist.
}
\references{
\itemize{
    \item Wilson, James D., Palowitch, John, Bhamidi, Shankar, and Nobel, Andrew B. (2017) "Significance based 
    extraction in multilayer networks with heterogeneous community structure." Journal of Machine Learning Research
}
}
\author{
James D. Wilson
}
\keyword{community}
\keyword{configuration}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{models}
\keyword{multilayer}
\keyword{networks,}
\keyword{random}
