% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_CM.R
\name{expected.CM}
\alias{expected.CM}
\title{expected.CM}
\usage{
expected.CM(adjacency)
}
\arguments{
\item{adjacency:}{a list object whose tth entry is an adjacency matrix representing the tth layer of a multilayer network.}
}
\value{
\itemize{
     \item P: a list object whose tth entry is the expected adjacency matrix of the tth layer 
}
}
\description{
Function that calculates the expected edge weight between each pair of nodes in each layer of a multilayer network.
}
\author{
James D. Wilson
}
\references{
\itemize{
    \item Wilson, James D., Palowitch, John, Bhamidi, Shankar, and Nobel, Andrew B. (2016) "Significance based 
    extraction in multilayer networks with heterogeneous community structure."
}
}
\keyword{community}
\keyword{configuration}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{models}
\keyword{multilayer}
\keyword{networks,}
\keyword{random}

