% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialization}
\alias{initialization}
\title{initialization}
\usage{
initialization(adjacency, prop.sample, m, n)
}
\arguments{
\item{adjacency:}{a list object whose tth entry is an adjacency matrix representing the tth layer of a multilayer network.}

\item{prop.sample:}{the proportion of vertices one would like to search over for initialization. Example: prop.sample = 0.05
specifies that one will obtain 0.05 * n randomly selected vertex neighborhoods for initialization, where n = number of nodes in each layer.}
}
\value{
\itemize{
     \item neighborhoods: a list object of length prop.sample * n, where each entry contains
     a vertex set and layer set from which multilayer.extraction can be run.
}
}
\description{
Function that generates randomly selected neighborhood vertex-layer sets to begin the 
multilayer.extraction algorithm.
}
\details{
A neighborhood of vertex u is defined as the collection of vertices that 
have higher than the mean connectivity of vertex u, when aggregated across layers. The 
chosen layer set is a random sample of size m/2.
}
\references{
\itemize{
    \item Wilson, James D., Palowitch, John, Bhamidi, Shankar, and Nobel, Andrew B. (2016) "Significance based 
    extraction in multilayer networks with heterogeneous community structure."
}
}
\author{
James D. Wilson
}
\keyword{community}
\keyword{configuration}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{models}
\keyword{multilayer}
\keyword{networks,}
\keyword{random}
