% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer_extraction.R
\name{multilayer.extraction}
\alias{multilayer.extraction}
\title{multilayer.extraction}
\usage{
multilayer.extraction(adjacency, seed = 123, min.score = 0,
  prop.sample = 0.05, directed = c(FALSE, TRUE))
}
\arguments{
\item{adjacency:}{a list object whose tth entry is an adjacency matrix representing the tth layer of a multilayer network.}

\item{seed:}{seed for reproducibility. The initial neighborhoods that act as seeds for the multilayer extraction algorithm
are random in this algorithm; hence, a seed will need to be set for reproducible results. Default is 123.}

\item{min.score:}{the minimum score allowable for an extracted community. Default is 0.}

\item{prop.sample:}{the proportion of vertices one would like to search over for initialization. Example: prop.sample = 0.05
specifies that one will obtain 0.05 * n randomly selected vertex neighborhoods for initialization, where n = number of nodes in each layer.
Default is 0.05.}
}
\value{
A MultilayerCommunity object, which is a list containing the following objects
\itemize{
     \item Community.List: a list of vertex-layer communities extracted from the algorithm
     \item Diagnostics: the diagnostics associated with each extracted community. This is a summary of 
     each community, and includes for each level of overlap parameter Beta the mean score, and the total number 
     of communities. This is used for determining the overall number of communities in a multilayer network. 
}
}
\description{
Function that identifies statistically significant vertex-layer communities in multilayer networks.
}
\references{
\itemize{
    \item Wilson, James D., Palowitch, John, Bhamidi, Shankar, and Nobel, Andrew B. (2017) "Significance based 
    extraction in multilayer networks with heterogeneous community structure." Journal of Machine Learning Research
}
}
\author{
James D. Wilson
}
\keyword{community}
\keyword{configuration}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{models}
\keyword{multilayer}
\keyword{networks,}
\keyword{random}
