% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity_matrix.R
\name{multilayer.modularity.matrix}
\alias{multilayer.modularity.matrix}
\title{multilayer.modularity.matrix}
\usage{
multilayer.modularity.matrix(adjacency, directed = FALSE)
}
\arguments{
\item{adjacency:}{an edgelist. Use adjacency.to.edgelist function to obtain this object}
}
\value{
mod.matrix: an igraph object representing the modularity matrix for each layer
}
\description{
Function that calculates the modularity matrix for each layer in a multilayer network
}
\references{
\itemize{
    \item Wilson, James D., Palowitch, John, Bhamidi, Shankar, and Nobel, Andrew B. (2017) "Significance based 
    extraction in multilayer networks with heterogeneous community structure." Journal of Machine Learning Research
}
}
\author{
James D. Wilson
}
\keyword{community}
\keyword{configuration}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{models}
\keyword{multilayer}
\keyword{networks,}
\keyword{random}
