% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MultilayerCommunity}
\alias{plot.MultilayerCommunity}
\title{plot.MultilayerCommunity}
\usage{
\method{plot}{MultilayerCommunity}(multilayer.object, main = "")
}
\arguments{
\item{multilayer.object:}{a MultilayerCommunity object that contains communities extracted from the multilayer.extraction algorithm.}

\item{main:}{the title of the plot. Optional. Default is "".}
}
\value{
A plot showing the total number of communities and average score of the communities across a grid of overlap parameters beta.
}
\description{
Function plots the average score and total number of communities extracted across a grid of overlap parameters beta.
}
\details{
This is a useful diagnostic plot to determine how many communities were extracted in the observed multilayer network. At beta = 0,
all of the communities are disjoint. As beta increases, the communities are allowed to overlap more. See the below reference for more details about this function.
}
\references{
\itemize{
    \item Wilson, James D., Palowitch, John, Bhamidi, Shankar, and Nobel, Andrew B. (2016) "Significance based 
    extraction in multilayer networks with heterogeneous community structure."
}
}
\author{
James D. Wilson
}
\keyword{community}
\keyword{configuration}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{models}
\keyword{multilayer}
\keyword{networks,}
\keyword{random}
