% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.R
\name{refine}
\alias{refine}
\title{refine}
\usage{
refine(Multilayer.object, k, m, n)
}
\arguments{
\item{Multilayer.object:}{a MultilayerCommunity object that contains communities extracted from the multilayer.extraction algorithm.}

\item{k:}{the number of communities that one would like to keep.}

\item{m:}{the total number of layers in the original multilayer network.}

\item{n:}{the total number of vertices in the original multilayer network.}
}
\value{
A list that contains the following objects
\itemize{
     \item Layers: an m x k matrix whose (i,j)th entry is non-zero if layer i is contained
     within community j. The value of the (i,j)th entry is the score of that community.
     \item Vertices: an n x k matrix whose (i,j)th entry is 1 if vertex i is contained within community j.
}
}
\description{
Function that identifies statistically significant vertex-layer communities in multilayer networks.
}
\references{
\itemize{
    \item Wilson, James D., Palowitch, John, Bhamidi, Shankar, and Nobel, Andrew B. (2016) "Significance based 
    extraction in multilayer networks with heterogeneous community structure."
}
}
\author{
James D. Wilson
}
\keyword{community}
\keyword{configuration}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{models}
\keyword{multilayer}
\keyword{networks,}
\keyword{random}
