% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{score}
\usage{
score(mod.matrix, vertex.set, layer.set, n)
}
\arguments{
\item{adjacency:}{a list object whose tth entry is an adjacency matrix representing the tth layer of a multilayer network.}

\item{expected:}{the expected value of the connectivity in the multilayer network adjacency. This can be calculated using the 
expected.CM function in this package.}

\item{vertex.set:}{a numeric specifying the nodes within the multilayer community of interest.}

\item{layer.set:}{a numeric specifying the layers within the multilayer community of interest.}
}
\value{
\itemize{
     \item score: the score of the multilayer community
     }
}
\description{
Function that calculates the score of a multilayer vertex - layer community
}
\references{
\itemize{
    \item Wilson, James D., Palowitch, John, Bhamidi, Shankar, and Nobel, Andrew B. (2016) "Significance based 
    extraction in multilayer networks with heterogeneous community structure."
}
}
\author{
James D. Wilson
}
\keyword{community}
\keyword{configuration}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{models}
\keyword{multilayer}
\keyword{networks,}
\keyword{random}
